Moved card & card-big components to features/games
