--- frontend/components/CardPlatform/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CardPlatform'
