+++ frontend/features/games/components/GameCard/Platform/Platform.module.scss
