+++ frontend/features/games/components/GameCard/index.ts
@@ -1,1 +1,2 @@
 export { GameCard } from './GameCard'
+export { GameCardBig } from './GameCardBig'
